TST: Add tests/fixtures/csv.py
- and move `specific_csv_path` into it,
  from `fixtures/dataframe`
- rename to `specific_prep_csv_path`
- also add `dummy_tmpfile_csv` to use in `test_cli/test_prep`
- use `specific_prep_csv_path` in fixtures/dataframe.py
- fix usage of `specific_prep_csv_path` in test_csv.py
