+++ tests/fixtures/__init__.py
@@ -1,6 +1,7 @@
 from .annot import *
 from .audio import *
 from .config import *
+from .csv import *
 from .dataframe import *
 from .device import *
 from .model import *
