+++ tests/test_csv.py
@@ -18,12 +18,12 @@ def test_has_unlabeled(config_type,
                        expected_result,
                        model,
                        specific_config_toml,
-                       specific_csv_path):
-    csv_path = specific_csv_path(config_type,
-                                 model,
-                                 annot_format,
-                                 audio_format,
-                                 spect_format)
+                       specific_prep_csv_path):
+    csv_path = specific_prep_csv_path(config_type,
+                                      model,
+                                      annot_format,
+                                      audio_format,
+                                      spect_format)
     config_toml = specific_config_toml(config_type,
                                        model,
                                        annot_format,
